Change Eth2 references in copy to LYXt
Noticed a lot of references/lines of copy that are only relevant to Eth, some could/should be rewritten or removed. Left a few as Eth as I was unsure.
Noted minor changes to give a start. If there is a guide or other documentation for this - I can help further.